feat(BaseService): support stream=True in BaseService.send()
Fixes: arf/planning-sdk-squad/issues/901

This change allows generated operation code to pass 'stream=True' to
the BaseService.send() method, which will in turn pass it on to the
underlying requests.request() method.  The net result is that
the response body (even if it is a JSON object) will be "streamed",
which means it will be made available via the requests.Response object
rather than consumed and set on the DetailedResponse object.
To access the streamed response, simply retrieve the requests.Response
object by calling DetailedResponse.get_result(), then use the "iter_content"
method on the returned requests.Response instance.