fix: use millisecond precision when marshalling date-time values
Fixes arf/planning-sdk-squad#2313

This commit makes a minor change to the datetime_to_string function
so that it uses millisecond precision when converting a datetime value
to a string.  This is aligned with the API handbook requiremetns re:
the wire format for date-time values.
I also added a bunch of tests that are (more or less) similar to
the set of tests in the Java and Go cores.