fix: silently ignore missing configuration file
While attempting to read configuration properties from
the config file specified via ${IBM_CREDENTIALS_FILE}, the
read_external_sources() function should silently ignore
an error if the file doesn't exist, and then proceed
to the next step of the config algorithm.
This was not working correctly, but this commit fixes that.