fix: multiple IAM based authenticator overrides (#124)
This commit makes the `request_payload` instance variable. Previously it was a class variable and that could cause issues if there were more than one IAM based authenticator in the same project.
There are also some unicode string and utf-8 encoding related changes to make the code compatible with the latest pylint release.