fix: detect JSON unmarshal errors in responses
This commit fixes BaseService.send() so that it
will now  detect a JSON parsing error when processing
a success (2xx) response.
If, in fact, a JSON parsing error occurs while
trying to process a JSON response body,
then a requests.exceptions.JSONDecodeError
is raised and propagated back to the caller.

Signed-off-by: Phil Adams <phil_adams@us.ibm.com>