fix: add correct support for compressing file-like objects (#174)
This commit contains the fix for a bug in the request preparation step.
Previously file-like objects couldn't be compressed and the code
threw an exception if the user tried to do it. The solution in this commit
is a helper class, called `GzipStream`. It acts like a "middleware"
between the reader (requests/urrlib3) and the writer (gzip) and
compresses the data on the fly.

Signed-off-by: Norbert Biczo <pyrooka@users.noreply.github.com>