fix: use on-the-fly compression only for file objects
When we introduced the new `GzipStream` utility class
which handles the on-the-fly compression of the file
like object. We also started using it for in-memory
objects, like bytes, but it has introduced a minor
incompatibility with the existing SDKs.
This commit introduces a fix for that, by using the
new utility class only for file-like objects and handling
the other types as before (compressing in memory).

Signed-off-by: Norbert Biczo <pyrooka@users.noreply.github.com>