feat: rename `ApiException.code` to `ApiException.status_code` (#185)
This commit fixes a small inconsistency in the naming
of the HTTP status code attributes between the successful
and the error responses. From now on, the right way to obtain
the status code from the `ApiException` is, using the `status_code`
attribute. `code` will remain for some time, but now it's deprecated
and it will be removed in the near future.

Signed-off-by: Norbert Biczo <pyrooka@users.noreply.github.com>