fix: improve the detection and loading of default certificates (#197)
Turned out in some cases - especially in containers - the certificate verification is still failing due to the missing certs. It's because the location of those files really depend on the system and the OpenSSL configuration.
This commit adds a workaround for this problem, by loading the default certs from the location that the `certifi` package reports. The `requests` package uses the same logic, so it should cause no issues.

Signed-off-by: Norbert Biczo <pyrooka@users.noreply.github.com>