fix: remove test command and related code from `setup.py` (#203)
The `test` command and related functions were deprecated for 5 years
and it seems the time has come: setuptools maintainers removed it
yesterday and that move has broken a bunch of Python packages
including this. This commit removes all references to that command
and any other code related to that. That means there will be no
more `python setup.py test`, but it wasn't be used for a long time.

Signed-off-by: Norbert Biczo <pyrooka@users.noreply.github.com>